pyproject.toml
[project]
name = "airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.0.1",
    "torch>=2.7.1",
    "transformers>=4.52.4",
]
